extracts the createBook function to lib
