+++ src/services/bookService.js
@@ -2,6 +2,7 @@ import db from '../models';
 import { isbnValidation } from '../helpers/validationHelper';
 import axios from 'axios';
 import Sequelize from 'sequelize';
+import createBooks from '../lib/createBooks';
 
 const { Op } = Sequelize;
 
@@ -243,39 +244,6 @@ exports.addBook = async req => {
             msg: 'Book added.',
             data
         };
-    } catch (error) {
-        return {
-            error: true,
-            statusCode: 500,
-            error
-        };
-    }
-};
-
-const createBooks = async bookData => {
-    try {
-        let newBooks = [];
-
-        // Creates local DB record for each book returned from gbBooks
-        for (let i = 0; i < bookData.length; i++) {
-            // checks if the book record exists in our database by isbn
-            const book = await db.book.findOne({
-                where: {
-                    isbn: bookData[i].isbn
-                }
-            });
-
-            // if it doesn't, go ahead and create that new record
-            if (!book) {
-                const newBook = await db.book.create(bookData[i]);
-
-                // push the new record to an array
-                newBooks.push(newBook.dataValues);
-            }
-        }
-
-        // return all new book records
-        return newBooks;
     } catch (error) {
         return {
             error: true,
